fix: transacttion manager

+++ src/schemas/requests/billing.py
@@ -3,6 +3,6 @@ from pydantic import BaseModel
 
 
 class TopUpBillingRequest(BaseModel):
-    atl_amount: int
+    atl_amount: float
     access_token: str
     invoice_id: str
